+++ server/server.js
@@ -3,7 +3,7 @@ var express = require('express');
 //bcrypt
 var bodyParser = require('body-parser');
 //express session
-var handlers = require('./handlers');
+var handlers = require('./ORMhandlers');
 var app = express(); 
 
 app.use(bodyParser.json());
